card-words-ai/pyproject.toml
[tool.poetry]
name = "card-words-ai"
version = "0.1.0"
description = "AI-powered vocabulary review system using XGBoost"
authors = ["Card Words Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.4.0"
pydantic-settings = "^2.0.0"
xgboost = "^2.0.0"
scikit-learn = "^1.3.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
sqlalchemy = "^2.0.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.0"
redis = "^5.0.0"
python-dotenv = "^1.0.0"
joblib = "^1.3.0"
structlog = "^23.2.0"
pyjwt = "^2.8.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.10.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
